Add PlotMemory to own file, and make it public
